+++ klutch/klutch.go
@@ -15,39 +15,42 @@ import (
 )
 
 const (
-	demoTitle               = "Klutch Demo"
-	mgmtClusterInfoFilePath = "klutch"
-	mgmtClusterInfoFileName = "management-cluster-info.yaml"
-	mgmtClusterName         = "klutch-management"
-	consumerClusterName     = "klutch-consumer"
-	contextMgmt             = "kind-" + mgmtClusterName
-	contextConsumer         = "kind-" + consumerClusterName
+	demoTitle                       = "Klutch Demo"
+	controlPlaneClusterInfoFilePath = "klutch"
+	controlPlaneClusterInfoFileName = "control-plane-cluster-info.yaml"
+	controlPlaneClusterName         = "klutch-control-plane"
+	appClusterName                  = "klutch-app"
+	contextControlPlane             = "kind-" + controlPlaneClusterName
+	contextApp                      = "kind-" + appClusterName
 )
 
 var (
 	PortFlag int = 8080
 )
 
-// ManagementClusterInfo contains information about the created management cluster.
-type ManagementClusterInfo struct {
+// ControlPlaneClusterInfo contains information about the created Control Plane Cluster.
+type ControlPlaneClusterInfo struct {
 	Host        string `yaml:"host"`
 	IngressPort string `yaml:"ingressPort"`
 }
 
 type KlutchManager struct {
-	mgmtK8s     *k8s.KubeClient
-	consumerK8s *k8s.KubeClient
+	// cpK8s is the Kubernetes client for the Klutch Control Plane Cluster.
+	cpK8s *k8s.KubeClient
+
+	// appK8s is the Kubernetes client for the Klutch App Cluster.
+	appK8s *k8s.KubeClient
 }
 
 func NewKlutchManager() *KlutchManager {
 	return &KlutchManager{
-		mgmtK8s:     k8s.NewKubeClient(contextMgmt),
-		consumerK8s: k8s.NewKubeClient(contextConsumer),
+		cpK8s:  k8s.NewKubeClient(contextControlPlane),
+		appK8s: k8s.NewKubeClient(contextApp),
 	}
 }
 
-// DeployKlutchClusters deploys the central management cluster with all Klutch components, and a
-// consumer cluster.
+// DeployKlutchClusters deploys the Control Plane Cluster with all Klutch components, and a
+// app cluster.
 func DeployKlutchClusters() {
 	makeup.PrintWelcomeScreen(
 		demo.UnattendedMode,
@@ -62,23 +65,23 @@ func DeployKlutchClusters() {
 	checkDeployPrerequisites()
 
 	klutch := NewKlutchManager()
-	klutch.deployCentralManagementCluster()
-	klutch.deployConsumerCluster()
+	klutch.deployControlPlaneCluster()
+	klutch.deployAppCluster()
 	printSummary()
 }
 
-func (k *KlutchManager) deployCentralManagementCluster() {
+func (k *KlutchManager) deployControlPlaneCluster() {
 	hostIP, err := determineHostLocalIP()
 	if err != nil {
 		makeup.ExitDueToFatalError(err, "Couldn't obtain the host's local IP address. Aborting...")
 	}
 
-	makeup.PrintInfo(fmt.Sprintf("Using IP address `%s` for the central management cluster.", hostIP))
+	makeup.PrintInfo(fmt.Sprintf("Using IP address `%s` for the Control Plane Cluster.", hostIP))
 
 	port := strconv.Itoa(PortFlag)
-	DeployManagementKindCluster(mgmtClusterName, hostIP, port)
-	WaitForKindCluster(k.mgmtK8s)
-	writeManagementClusterInfoToFile(demo.DemoConfig.WorkingDir, hostIP, port)
+	DeployControlPlaneKindCluster(controlPlaneClusterName, hostIP, port)
+	WaitForKindCluster(k.cpK8s)
+	writeControlPlaneClusterInfoToFile(demo.DemoConfig.WorkingDir, hostIP, port)
 
 	k.DeployIngressNginx()
 	k.WaitForIngressNginx()
@@ -93,56 +96,56 @@ func (k *KlutchManager) deployCentralManagementCluster() {
 
 	makeup.H1("Klutch components are deployed. Deploying the a8s stack...")
 	makeup.WaitForUser(demo.UnattendedMode)
-	a8s := demo.NewA8sDemoManager(contextMgmt)
+	a8s := demo.NewA8sDemoManager(contextControlPlane)
 	a8s.DeployA8sStack()
 }
 
-func (k *KlutchManager) deployConsumerCluster() {
+func (k *KlutchManager) deployAppCluster() {
 
-	DeployConsumerCluster(consumerClusterName)
-	WaitForKindCluster(k.consumerK8s)
+	DeployAppCluster(appClusterName)
+	WaitForKindCluster(k.appK8s)
 
-	switchContext(contextConsumer) // in case the consumer cluster already existed, switch to it.
+	switchContext(contextApp) // in case the app cluster already existed, switch to it.
 }
 
 func printSummary() {
 	makeup.PrintH1("Summary")
 	makeup.Print("You've successfully accomplished the followings steps:")
-	makeup.PrintCheckmark("Deployed a Klutch management Kind cluster.")
+	makeup.PrintCheckmark("Deployed a Klutch Control Plane Cluster with Kind.")
 	makeup.PrintCheckmark("Deployed Dex Idp and the anynines klutch-bind backend.")
 	makeup.PrintCheckmark("Deployed Crossplane and the Kubernetes provider.")
 	makeup.PrintCheckmark("Deployed the Klutch Crossplane configuration package.")
-	makeup.PrintCheckmark("Deployed Klutch API Service Export Templates to make the Klutch Crossplane APIs available to consumer clusters.")
+	makeup.PrintCheckmark("Deployed Klutch API Service Export Templates to make the Klutch Crossplane APIs available to App Clusters.")
 	makeup.PrintCheckmark("Deployed the a8s Stack.")
-	makeup.PrintCheckmark("Deployed a consumer cluster.")
-	makeup.PrintSuccessSummary("You are now ready to bind APIs from the consumer cluster using the `a9s klutch bind` command.")
+	makeup.PrintCheckmark("Deployed an App Cluster.")
+	makeup.PrintSuccessSummary("You are now ready to bind APIs from the App Cluster using the `a9s klutch bind` command.")
 }
 
-// Writes information about the management cluster to a file, to give other commands such as `bind` the information they need.
-func writeManagementClusterInfoToFile(workDir string, hostIP string, ingressPort string) {
-	info := &ManagementClusterInfo{
+// Writes information about the Control Plane Cluster to a file, to give other commands such as `bind` the information they need.
+func writeControlPlaneClusterInfoToFile(workDir string, hostIP string, ingressPort string) {
+	info := &ControlPlaneClusterInfo{
 		Host:        hostIP,
 		IngressPort: ingressPort,
 	}
 
 	data, err := yaml.Marshal(info)
 	if err != nil {
-		makeup.ExitDueToFatalError(err, "Unexpected error while writing managment cluster info to file.")
+		makeup.ExitDueToFatalError(err, "Unexpected error while writing Control Plane Cluster info to file.")
 	}
 
-	path := filepath.Join(workDir, mgmtClusterInfoFilePath)
+	path := filepath.Join(workDir, controlPlaneClusterInfoFilePath)
 	err = os.MkdirAll(path, 0755)
 	if err != nil {
-		makeup.ExitDueToFatalError(err, fmt.Sprintf("Unexpected error while writing managment cluster info to file. Could not create path %s", path))
+		makeup.ExitDueToFatalError(err, fmt.Sprintf("Unexpected error while writing Control Plane Cluster info to file. Could not create path %s", path))
 	}
 
-	file := filepath.Join(path, mgmtClusterInfoFileName)
+	file := filepath.Join(path, controlPlaneClusterInfoFileName)
 	err = os.WriteFile(file, data, 0644)
 	if err != nil {
-		makeup.ExitDueToFatalError(err, fmt.Sprintf("Unexpected error while writing managment cluster info to file. Could not write to file %s", file))
+		makeup.ExitDueToFatalError(err, fmt.Sprintf("Unexpected error while writing Control Plane Cluster info to file. Could not write to file %s", file))
 	}
 
-	makeup.PrintInfo(fmt.Sprintf("Wrote management cluster information to %s", file))
+	makeup.PrintInfo(fmt.Sprintf("Wrote Control Plane Cluster information to %s", file))
 }
 
 // Checks if prerequisites of the deploy command are met.